# Add `RuleRunner` as a Pytest-style replacement to `TestBase`