--- a/src/python/pants/engine/internals/uuid_test.py
+++ b/src/python/pants/engine/internals/uuid_test.py
@@ -3,23 +3,26 @@
 
 from uuid import UUID
 
+import pytest
+
 from pants.engine.internals.uuid import UUIDRequest
 from pants.engine.internals.uuid import rules as uuid_rules
 from pants.engine.rules import QueryRule
-from pants.testutil.test_base import TestBase
+from pants.testutil.rule_runner import RuleRunner
 
 
-class UUIDTest(TestBase):
-    @classmethod
-    def rules(cls):
-        return (*super().rules(), *uuid_rules(), QueryRule(UUID, (UUIDRequest,)))
+@pytest.fixture
+def rule_runner() -> RuleRunner:
+    return RuleRunner(rules=[*uuid_rules(), QueryRule(UUID, (UUIDRequest,))])
 
-    def test_distinct_uuids(self):
-        uuid1 = self.request_product(UUID, [UUIDRequest()])
-        uuid2 = self.request_product(UUID, [UUIDRequest()])
-        assert uuid1 != uuid2
 
-    def test_identical_uuids(self):
-        uuid1 = self.request_product(UUID, [UUIDRequest(randomizer=0.0)])
-        uuid2 = self.request_product(UUID, [UUIDRequest(randomizer=0.0)])
-        assert uuid1 == uuid2
+def test_distinct_uuids(rule_runner: RuleRunner) -> None:
+    uuid1 = rule_runner.request_product(UUID, [UUIDRequest()])
+    uuid2 = rule_runner.request_product(UUID, [UUIDRequest()])
+    assert uuid1 != uuid2
+
+
+def test_identical_uuids(rule_runner: RuleRunner) -> None:
+    uuid1 = rule_runner.request_product(UUID, [UUIDRequest(randomizer=0.0)])
+    uuid2 = rule_runner.request_product(UUID, [UUIDRequest(randomizer=0.0)])
+    assert uuid1 == uuid2