… src/python/pants/backend/project_info/list_roots_test.py
… src/python/pants/backend/python/pants_requirement_test.py
… src/python/pants/backend/python/pipenv_requirements_test.py
… src/python/pants/backend/python/python_requirements_test.py
… src/python/pants/core/util_rules/archive_test.py
… src/python/pants/core/util_rules/filter_empty_sources_test.py
… src/python/pants/core/util_rules/source_files_test.py
… src/python/pants/core/util_rules/stripped_source_files_test.py
… src/python/pants/engine/internals/build_files_test.py
… src/python/pants/engine/platform_test.py
… src/python/pants/source/filespec_test.py
… src/python/pants/source/source_root_test.py